// aiot-frontend/src/api/generated/device-service/models/device-v1-device.ts
/* tslint:disable */
/* eslint-disable */
/**
 * 
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.0.1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */



/**
 * Device message
 */
export interface DeviceV1Device {
    'id'?: string;
    'deviceKey'?: string;
    'name'?: string;
    'productId'?: string;
    'productKey'?: string;
    'productName'?: string;
    'state'?: string;
    'enabled'?: boolean;
    'lastOnlineTime'?: string;
    'metadata'?: string;
    'createdAt'?: string;
    'updatedAt'?: string;
    'deletedAt'?: string;
    'lastOfflineTime'?: string;
}